% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTRSpace.R
\name{create_rs_simago}
\alias{create_rs_simago}
\title{Create Scimago Research Space}
\usage{
create_rs_simago()
}
\arguments{
\item{data}{A numeric matrix with entities \eqn{i} in the rows and categories \eqn{j} in the columns. Cells show the respective value (value of abundance) of entity \eqn{i} in the category \eqn{j}. It can also be a transpose of the previous matrix, that is, a matrix with categories in the rows and entities in the columns. Yet in that case, the argument "category_row" has to be set to TRUE. The matrix must include names for the rows and the columns. The argument "data", also accepts a dataframe with three columns in the following order: entity, category and value.}

\item{category_row}{A flag to indicate that categories are in the rows. The analysis assumes that the categories are in the columns of the matrix. If the categories are in the rows and the entities in the columns, then the argument "category_row" has to be set to TRUE. The default value is FALSE.}
}
\value{
An iGraph object.
}
\description{
Open RSpace on Scimago taxonomy, plots and returns an iGraph object.
}
\examples{
create_rs_scimago()
}
