% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ignore_utils.R
\name{create_scimago_taxo}
\alias{create_scimago_taxo}
\title{ignore, just used to build data
create data for Scimago Taxonomy}
\usage{
create_scimago_taxo()
}
\description{
ignore, just used to build data
create data for Scimago Taxonomy
}

