% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_benchmark_map}
\alias{get_benchmark_map}
\title{Get a benchmark map}
\usage{
get_benchmark_map(layout = "fr", rs = NULL, pl = FALSE, lab = FALSE,
  pl_mst = FALSE, mean_degree = 4, cex = 1, mst = FALSE,
  pl_seed = pl_seed)
}
\arguments{
\item{data_raw}{a dataframe with production data use function load_raw_data()}

\item{init}{an integer pointing the initial year}

\item{end}{an integer pointing the final year}

\item{producer}{name of the column in data_raw that contains the name or id of producers. It should be factor.}

\item{category}{name of the column in data_raw that contains the name or id of categories. It should be factor.}

\item{value}{name of the column in data_raw that contains the feature to aggregate. should be numeric}

\item{agg}{function to aggregate, usually sum or mean. If NaN, no aggregation is computed}
}
\value{
a Data Frame to overlay.
}
\description{
this function returns an igraph object based on links or in .Rdata created previously
}
\examples{
load_data_interval()
}
