% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{get_data_to_overlay}
\alias{get_data_to_overlay}
\title{Get data to overlay}
\usage{
get_data_to_overlay(data_interval, n = 10, init = NaN, end = NaN,
  by = 1, cum = FALSE, min_prod = 1, year = "year", producer, category,
  value, agg_fun = mean, list_prod = NaN)
}
\arguments{
\item{n}{Number of samples to use. -1 for the complete number of producers.}

\item{init}{an integer pointing the initial year}

\item{end}{an integer pointing the final year}

\item{by}{an integer describing the time window}

\item{cum}{if FALSE it uses mobile time windows, otherwise it uses allways the same init year.}

\item{min_prod}{minimum production to filter producers}

\item{agg_fun}{function to aggregate, usually sum,}

\item{list_prod}{default NaN. A finite list of producers}

\item{what_agg}{name of column to aggregate 'wgh_jfrac',}
}
\value{
a Data Frame to overlay.
}
\description{
this function read data in intervals, aggregate it and computes measures of normalization of the data as RCA. Also subset and filter per a list or value of producers
}
\examples{
get_data_to_overlay()
}
