% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_intervals}
\alias{get_intervals}
\title{Get a dataframe with intervals}
\usage{
get_intervals(years, year_ini, year_fin, win = -1, decade = FALSE)
}
\arguments{
\item{years}{Vector of years in raw data}

\item{year_ini}{Initital year}

\item{year_fin}{Final year}

\item{win}{time window}

\item{decade}{deprected}
}
\value{
a Data Frame with year start, year end and interval name.
}
\description{
this function computes intervals according period of time
}
\examples{
get_intervals()
}
