% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{get_prod_name}
\alias{get_prod_name}
\title{Get the name of a producer}
\usage{
get_prod_name(id_prod, producers)
}
\arguments{
\item{id_prod}{an identifier for a producer.}

\item{producers}{a dataframe with columns information of producers, it must contain column id and ideally name.}
}
\value{
a name for the producer, based on its id. It returns the id in case the column name was not found.
}
\description{
this function returns the name of a producer based on an id defined by id_prod.
}
\examples{
get_prod_name()
}
