% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_data_interval}
\alias{load_data_interval}
\title{Load data intervals}
\usage{
load_data_interval(data_raw = load_raw_data(), init = NaN, end = NaN,
  year = "year", producer, category, value = "value", agg = "mean")
}
\arguments{
\item{data_raw}{a dataframe with production data use function load_raw_data()}

\item{init}{an integer pointing the initial year}

\item{end}{an integer pointing the final year}

\item{producer}{name of the column in data_raw that contains the name or id of producers. It should be factor.}

\item{category}{name of the column in data_raw that contains the name or id of categories. It should be factor.}

\item{value}{name of the column in data_raw that contains the feature to aggregate. should be numeric}

\item{agg}{function to aggregate, usually sum or mean. If NaN, no aggregation is computed}
}
\value{
a Data Frame to overlay.
}
\description{
this function look into data_raw and then subset by intervals aggregating information
}
\examples{
load_data_interval()
}
