% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_raw_data}
\alias{load_raw_data}
\title{Load Raw Data}
\usage{
load_raw_data(n = -1, path_raw_files = "data_raw", col_classes = NaN,
  sep = NaN)
}
\arguments{
\item{n}{Number of lines to use- -1 implies all the lines}

\item{path_raw_files}{A path to local files where the files with data are located}

\item{col_classes}{A vector with configuration of what column to read and of what type. Use factor for IDs or labels and NA for default numeric values}

\item{sep}{A separator for columns in raw files. If NaN, an automatic process will be conducted to detect the separator.}
}
\value{
a Data Frame to overlay.
}
\description{
this function read a folder and import raw data to a dataframe. It automatically finds separator.
}
\examples{
load_data()
}
