% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{overlay_data_3}
\alias{overlay_data_3}
\title{Get data to overlay}
\usage{
overlay_data_3(data_to_use, nodes, producers = NULL, what_eval = "share",
  min_to_grw = -1, min_to_dev = -1, pl_base = TRUE, pl_base_mst = TRUE,
  pl = TRUE, pl_roc = TRUE, pl_min_dens = 0.15, pl_node_sc = 2.5,
  pl_min_size = 0.7, pl_num_lab = 5, rs_to_eval = NULL,
  rs_mean_degree = 4, cex = 1, pl_w_pdf = FALSE, pl_seed = 69,
  dens_links = "all")
}
\arguments{
\item{data_to_use}{Data computed previously with get_data_to_overlay().}

\item{nodes}{a data frame with all information of nodes. It must be the same used by rs_to_eval}

\item{what_eval}{name of the column of the dataframe data_to_use to use in the evaluation}

\item{min_to_grw}{minimum value to define the stat growing}

\item{min_to_dev}{minimum value to define the state of developing}

\item{pl_base}{plot}

\item{pl_base_mst}{plot base with mst}

\item{pl}{To plot or not to plot,}

\item{pl_roc}{to plot or not to plot roc curve}

\item{rs_to_eval}{An iGraph to overlay on data}

\item{dens_links}{dens_links Which links to use in the evaluation of density.  'all' 'bench' just links in benchmark, 'filter' for filtered links}

\item{produces}{a data frame with information of producers. It must include columns id and name}

\item{pl_min_size=0.7, }{}
}
\value{
a Data Frame to overlay.
}
\description{
this function read data in intervals, aggregate it and computes measures of normalization of the data as RCA. Also subset and filter per a list or value of producers
}
\examples{
overlay_data_3()
}
