% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pantheon}
\alias{pantheon}
\title{SCImago taxonomy}
\format{A dataframe with nodes:
\describe{
\item{Country}{Name of the country}
\item{Occupation}{Occupation according to the taxonomy of Pantheon}
\item{Value}{ Quantity of globally famous people that were born in that country}
}}
\source{
Raw data before the aggregation was queried from
\url{http://www.scimagojr.com/} in 2016.
}
\usage{
pantheon
}
\description{
SCImago taxonomy
}
\examples{
data(scimago)
str(scimago)
summary(scimago)
pantheon[pantheon$Country=="Chile",]
}
\references{
SCImago. (2007). SJR-SCImago Journal & Country Rank.
}
\keyword{dataset}

