% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{plot_rocs}
\alias{plot_rocs}
\title{Evaluate roc curves}
\usage{
plot_rocs(data_eval_roc, producers = NULL, pl = FALSE)
}
\arguments{
\item{data_eval_roc}{a data frame processed previously with plot roc}

\item{producers}{a data frame with producers}
}
\value{
a Data Frame to overlay.
}
\description{
evaluate roc curves for each producer
}
\examples{
plot_rocs()
}
