% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResearchSpace.R
\name{plot_rs}
\alias{plot_rs}
\title{Plot Research Space}
\usage{
plot_rs(map = NULL, mean_degree = 4, mst = TRUE, pl_seed = 1,
  prop_to_lab = 0.2, cex = 1, pl = FALSE, pl_mst = FALSE)
}
\arguments{
\item{mean_degree}{a number of degree to filter links}

\item{mst}{if TRUE it adds links of the Minimum Spanning Tree}

\item{pl_seed}{the seed to plot}

\item{prop_to_lab}{proportion to label a vertex}

\item{pl}{if False it will not plot the map}

\item{pl_mst}{to plot or not to plot the MST alone}

\item{path_rs}{a path to the file containing the research space}
}
\value{
an iGraph.
}
\description{
this function takes a MATRIX OF ADJACENCY of a research space previously created with function create_rs() in order to filter it and plot it Use i.e:rs_sim_pr_1987_2014_n_-1_aw_0_jf_0_awjf_0.75_min_prod_0, rs_sim_pr_2000_2009_n_-1_aw_0_jf_0_awjf_0.75_min_prod_0
}
\examples{
plot_rs()
}
